kis병합

+++ Tlink/src/main/webapp/resources/js/work/comment.js
@@ -3,6 +3,7 @@
 
 
 $(`.commentInsert`).on("click", function(){
+    const commentCon=$(`.commentInsertContent`).text();
     const data={ 
         "commentCon"  : commentCon, 
         "workNo"      : workNo, 
@@ -10,25 +11,9 @@ $(`.commentInsert`).on("click", function(){
     };
     fetch("/commentWork", { method: "POST", headers: {"Content-Type" : "application/json"}, body: JSON.stringify(data) })
     .then (rep => rep.json())
-    .then (res => {
-
-
-        console.log(res);
-
-
-        // const work=res;
-        // work.workNo!=0 ?  alert("성공하였습니다.") : alert("실패하였습니다."); 
-        // if(work.workNo!=0){ 
-            
-        // $(this).parents(`td`).find(`ul`).append(`<li>
-        //         <span class="workNo" hidden>\${work.workNo}</span>
-        //         <a href="/workSheet?workNo=\${work.workNo}&projectNo=\${projectNo}"><span class="material-symbols-outlined">draft</span></a>
-        //         <span class="workTitle" contenteditable="true">\${work.workTitle}</span>
-        //         <span class="material-symbols-outlined deleteWork">close</span>
-        //     </li>`)
-        // }
-
-
+    .then (res => { console.log(res); 
+        $(`.commentList`).html("");
+        selectComment();
     })
     .catch(err => console.log(err))
         
@@ -38,57 +23,75 @@ $(`.commentInsert`).on("click", function(){
 
 
 
-$(`.commentDelete`).on("click", function(){
-    const commentNo=$(this).pFind(`section`,`.commentNo`).text();
-    console.log(commentNo);
-
-    const data={
-        "commentNo"  : commentNo, 
-    };
-    fetch("/commentWork", { method: "DELETE", headers: {"Content-Type" : "application/json"}, body: JSON.stringify(data) })
-    .then (rep => rep.json())
-    .then (res => {
-
-        console.log(res);
-
-    })
-    .catch(err => console.log(err))
-        
+$(document).on("click", `.commentDelete`, function(){
+    if(confirm("삭제하시겠습니까??")){
+        const commentNo=$(this).pFind(`section`,`.commentNo`).text();
+        console.log(commentNo);
+    
+        const data={
+            "commentNo"  : commentNo, 
+        };
+        fetch("/commentWork", { method: "DELETE", headers: {"Content-Type" : "application/json"}, body: JSON.stringify(data) })
+        .then (rep => rep.json())
+        .then (res => {
+            console.log(res);
+            $(`.commentList`).html("");
+            selectComment();
+        })
+        .catch(err => console.log(err))
+    }
+    return;        
 })
 
 
 
 function selectComment(){
     
-    console.log(workNo);
-
     const data={ 
         "workNo"      : workNo, 
     };
     fetch("/commentWork/select", { method: "POST", headers: {"Content-Type" : "application/json"}, body: JSON.stringify(data) })
     .then (rep => rep.json())
     .then (res => {
-
-
         console.log(res);
+        const arr=res;
+        for(let comment of arr){
+     
+            $(`.commentList`).append(
+                `<section>
+                    <span class="material-symbols-outlined">person_pin_circle</span>
+                    <span class="commentNo" hidden>${comment.commentNo}</span>
+                    <span class="userNo" hidden>${comment.userNo}</span>
+                    <span>${comment.userName}</span>
+                    ${userNo==comment.userNo ? '<sup class="commentDelete">x</sup>' : ''}
+                    <h6 class="commentContent" ${userNo==comment.userNo ? 'contenteditable="true"' : ''}>${comment.commentCon}</h6>
+                </section>`
+            );
 
-
-        // const work=res;
-        // work.workNo!=0 ?  alert("성공하였습니다.") : alert("실패하였습니다."); 
-        // if(work.workNo!=0){ 
-            
-        // $(this).parents(`td`).find(`ul`).append(`<li>
-        //         <span class="workNo" hidden>\${work.workNo}</span>
-        //         <a href="/workSheet?workNo=\${work.workNo}&projectNo=\${projectNo}"><span class="material-symbols-outlined">draft</span></a>
-        //         <span class="workTitle" contenteditable="true">\${work.workTitle}</span>
-        //         <span class="material-symbols-outlined deleteWork">close</span>
-        //     </li>`)
-        // }
-
-
+        }
     })
     .catch(err => console.log(err))
         
 }
 selectComment();
 
+
+
+
+$(document).on("focusout", `.commentContent`, function(){
+    const commentNo=$(this).pf(`section`,`.commentNo`).text();
+    const commentCon=$(this).text();
+    const data={ 
+        "commentNo"  : commentNo, 
+        "commentCon" : commentCon 
+    };
+    fetch("/commentWork", { method: "PUT", headers: {"Content-Type" : "application/json"}, body: JSON.stringify(data) })
+    .then (rep => rep.json())
+    .then (res => { 
+        console.log(res); 
+        $(`.commentList`).html("");
+        selectComment();
+    })
+    .catch(err => console.log(err))
+        
+})
